Memory & performance: remove silly profiles, let users define their own RAM allocations
The current system is broken (OpenCL available memory detection is non-standard-compliant and doesn't work), and even CPU path returns false-positives errors ("pipeline will run non-tiled BS even though BS memory detection thinks the buffer won't fit in memory).

Plus the code is cryptic and non-legible. Those performance profiles are non-flexible yet complicated.

Revert to a simple RAM usage definition for:
- OS/other apps headroom,
- mipmap cache (lighttable)
- pixelpipe cache (darkraoom).

Pixelpipe cache size takes the window size into account and gets refreshed when window is resized.

+++ src/common/mipmap_cache.h
@@ -93,7 +93,7 @@ typedef struct dt_mipmap_cache_t
   // real width and height are stored per element
   // (could be smaller than the max for this mip level,
   // due to aspect ratio)
-  uint32_t max_width[DT_MIPMAP_NONE], max_height[DT_MIPMAP_NONE];
+  size_t max_width[DT_MIPMAP_NONE], max_height[DT_MIPMAP_NONE];
   // size of an element inside buf
   size_t buffer_size[DT_MIPMAP_NONE];
